+++ src/Xunit.OpenCategories/AuthorAttribute.cs
@@ -1,31 +1,30 @@
-﻿namespace Xunit.OpenCategories;
+﻿using System;
+using Xunit.Sdk;
 
-/// <summary>
-/// Attribute to specify the author of a test class or method.
-/// </summary>
-/// <remarks>
-/// This attribute can be applied to both classes and methods, and it supports multiple usages.
-/// </remarks>
-[AttributeUsage(AttributeTargets.Assembly | AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = true)]
-public class AuthorAttribute : BaseAttribute
+namespace Xunit.OpenCategories
 {
     /// <summary>
-    /// Initializes a new instance of the <see cref="AuthorAttribute"/> class.
+    /// Attribute to specify the author of a test class or method.
     /// </summary>
-    /// <param name="authorName">The name of the author.</param>
-    public AuthorAttribute(string? authorName)
+    /// <remarks>
+    /// This attribute can be applied to both classes and methods, and it supports multiple usages.
+    /// </remarks>
+    [TraitDiscoverer(AuthorDiscoverer.DiscovererTypeName, DiscovererUtil.AssemblyName)]
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = true)]
+    public class AuthorAttribute : Attribute, ITraitAttribute
     {
-        AuthorName = authorName;
-    }
+        /// <summary>
+        /// Initializes a new instance of the <see cref="AuthorAttribute"/> class.
+        /// </summary>
+        /// <param name="authorName">The name of the author.</param>
+        public AuthorAttribute(string authorName)
+        {
+            AuthorName = authorName;
+        }
 
-    /// <summary>
-    /// Gets the name of the author.
-    /// </summary>
-    public string? AuthorName { get; }
-
-    /// <inheritdoc />
-    protected override void OptionalTraits(List<KeyValuePair<string, string>> traits)
-    {
-        AddOptionalTrait(traits,"Author", AuthorName);
+        /// <summary>
+        /// Gets the name of the author.
+        /// </summary>
+        public string AuthorName { get; }
     }
 }