+++ src/Xunit.OpenCategories/LocalTestAttribute.cs
@@ -1,53 +1,46 @@
-﻿namespace Xunit.OpenCategories;
+﻿using System;
+using Xunit.Sdk;
 
-/// <summary>
-/// For tests that should only be executed locally and excluded from automated pipeline runs.
-/// </summary>
-/// <example>
-/// Trying out LINQ for the first time, writing a piece of code to understand IEnumerable.Take and Skip.
-/// </example>
-[AttributeUsage(AttributeTargets.Assembly | AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = true)]
-public class LocalTestAttribute : BaseAttribute
+namespace Xunit.OpenCategories
 {
     /// <summary>
-    /// Initializes a new instance of the <see cref="LocalTestAttribute"/> class with a specified ID.
+    /// For tests that should only be executed locally and excluded from automated pipeline runs.
     /// </summary>
-    /// <param name="id">The ID associated with the local test.</param>
-    public LocalTestAttribute(string? id)
+    /// <example>
+    /// Trying out LINQ for the first time, writing a piece of code to understand IEnumerable.Take and Skip.
+    /// </example>
+    [TraitDiscoverer(LocalTestDiscoverer.DiscovererTypeName, DiscovererUtil.AssemblyName)]
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = true)]
+    public class LocalTestAttribute : Attribute, ITraitAttribute
     {
-        Id = id;
-    }
+        /// <summary>
+        /// Initializes a new instance of the <see cref="LocalTestAttribute"/> class with a specified ID.
+        /// </summary>
+        /// <param name="id">The ID associated with the local test.</param>
+        public LocalTestAttribute(string id)
+        {
+            Id = id;
+        }
 
-    /// <summary>
-    /// Initializes a new instance of the <see cref="LocalTestAttribute"/> class with a specified ID.
-    /// </summary>
-    /// <param name="id">The ID associated with the local test.</param>
-    public LocalTestAttribute(long id)
-    {
-        Id = id.ToString();
-    }
+        /// <summary>
+        /// Initializes a new instance of the <see cref="LocalTestAttribute"/> class with a specified ID.
+        /// </summary>
+        /// <param name="id">The ID associated with the local test.</param>
+        public LocalTestAttribute(long id)
+        {
+            Id = id.ToString();
+        }
 
-    /// <summary>
-    /// Initializes a new instance of the <see cref="LocalTestAttribute"/> class.
-    /// </summary>
-    public LocalTestAttribute()
-    {
-    }
+        /// <summary>
+        /// Initializes a new instance of the <see cref="LocalTestAttribute"/> class.
+        /// </summary>
+        public LocalTestAttribute()
+        {
+        }
 
-    /// <summary>
-    /// Gets the ID associated with the local test.
-    /// </summary>
-    public string? Id { get; } = string.Empty;
-
-    /// <inheritdoc />
-    protected override void OptionalTraits(List<KeyValuePair<string, string>> traits)
-    {
-        AddOptionalTrait(traits, "LocalTest", Id);
-    }
-
-    /// <inheritdoc />
-    protected override void MandatoryTraits(List<KeyValuePair<string, string>> traits)
-    {
-        AddCategory(traits, "LocalTest");
+        /// <summary>
+        /// Gets the ID associated with the local test.
+        /// </summary>
+        public string Id { get; }
     }
 }