+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.1.9] - 2025-04-21
+
 ### Fixed
 
 - Bug where `Components` attribute was not adding the correct traits
@@ -87,7 +89,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Modified to use custom GA workflows
 - Published to NuGet
 
-[unreleased]: https://github.com/baynezy/Xunit.OpenCategories/compare/2.1.0.8...HEAD
+[unreleased]: https://github.com/baynezy/Xunit.OpenCategories/compare/2.1.1.9...HEAD
+[2.1.1.9]: https://github.com/baynezy/Xunit.OpenCategories/compare/2.1.0.8...2.1.1.9
 [2.1.0.8]: https://github.com/baynezy/Xunit.OpenCategories/compare/2.0.0.7...2.1.0.8
 [2.0.0.7]: https://github.com/baynezy/Xunit.OpenCategories/compare/1.2.0.5...2.0.0.7
 [1.2.0.5]: https://github.com/baynezy/Xunit.OpenCategories/compare/1.1.0.4...1.2.0.5